+++ src/Data/Generator/Combinators.hs
@@ -1,8 +1,4 @@
-{-# LANGUAGE CPP #-}
 {-# LANGUAGE UndecidableInstances, TypeOperators, FlexibleContexts, MultiParamTypeClasses, FlexibleInstances, TypeFamilies #-}
-#if defined(__GLASGOW_HASKELL__) && __GLASGOW_HASKELL__ >= 702 && __GLASGOW_HASKELL__ < 710
-{-# LANGUAGE Trustworthy #-}
-#endif
 
 -----------------------------------------------------------------------------
 -- |
@@ -53,9 +49,7 @@ module Data.Generator.Combinators
 import Prelude hiding
   ( mapM_, any, all, elem, filter, concatMap, and, or
   , sum, product, notElem, replicate, cycle, repeat
-#if __GLASGOW_HASKELL__ >= 710
   , foldMap
-#endif
   )
 import Control.Applicative (Alternative)
 import Control.Monad (MonadPlus)
@@ -66,11 +60,6 @@ import Data.Semigroup.Alternative (Alternate(..))
 import Data.Semigroup.Monad (Action(..))
 import Data.Semigroup.MonadPlus (MonadSum(..))
 import Data.Semigroup.Reducer (Reducer(..))
-
-#if !(MIN_VERSION_base(4,8,0))
-import Control.Applicative (Applicative)
-import Data.Monoid (Monoid(..))
-#endif
 
 -- | Efficiently 'mapReduce' a 'Generator' using the 'Traversal' monoid. A specialized version of its namesake from "Data.Foldable"
 --