+++ src/Data/Semigroup/Combinators.hs
@@ -1,7 +1,5 @@
-{-# LANGUAGE UndecidableInstances, TypeOperators, FlexibleContexts, MultiParamTypeClasses, FlexibleInstances, TypeFamilies, CPP #-}
-#if defined(__GLASGOW_HASKELL__) && __GLASGOW_HASKELL__ >= 702
+{-# LANGUAGE UndecidableInstances, TypeOperators, FlexibleContexts, MultiParamTypeClasses, FlexibleInstances, TypeFamilies #-}
 {-# LANGUAGE Trustworthy #-}
-#endif
 
 
 -----------------------------------------------------------------------------